openclaw-concierge/pyproject.toml
[project]
name = "openclaw-concierge"
version = "0.1.0"
description = "OpenClaw Concierge - Interview Phase + Setup Guide Generation"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.128.3",
    "uvicorn==0.40.0",
    "python-dotenv==1.2.1",
    "pydantic==2.11.7",
    "aiofiles==25.1.0",
    "httpx>=0.27.0",
]